move isDirectoryUsable to System
